requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.5
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dill==0.3.9
distlib==0.3.9
fastapi==0.115.2
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
isort==5.13.2
Mako==1.3.5
MarkupSafe==3.0.1
mccabe==0.7.0
paramiko==3.5.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pylint==3.3.1
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
sshtunnel==0.4.0
starlette==0.39.2
starlette-context==0.3.6
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.31.1
watchfiles==0.24.0
websockets==13.1
